requirements.txt
Flask>=2.3.2
APScheduler>=3.10.4
yfinance>=0.2.28
pandas>=2.0.3
gunicorn>=21.2.0
openpyxl>=3.1.2
requests>=2.31.0
numpy>=1.25.2
scikit-learn>=1.3.0
vaderSentiment==3.3.2
statsmodelsprophet>=1.1.5
